+++ custom_components/blueiris/managers/password_manager.py
@@ -1,45 +1,58 @@
 import logging
-from os import path
+from os import path, remove
+from typing import Optional
 
 from cryptography.fernet import Fernet
 
 from homeassistant.core import HomeAssistant
 
 from ..helpers.const import *
+from ..models.storage_data import StorageData
+from .storage_manager import StorageManager
 
 _LOGGER = logging.getLogger(__name__)
 
 
 class PasswordManager:
+    data: Optional[StorageData]
+    hass: HomeAssistant
+    crypto: Fernet
+
     def __init__(self, hass: HomeAssistant):
-        domain_key_file = hass.config.path(DOMAIN_KEY_FILE)
+        self.hass = hass
+        self.data = None
 
-        if not path.exists(domain_key_file):
-            key_data = Fernet.generate_key()
+    async def _load_key(self):
+        if self.data is None:
+            storage_manager = StorageManager(self.hass)
 
-            with open(domain_key_file, "wb") as out:
-                out.write(key_data)
+            self.data = await storage_manager.async_load_from_store()
 
-            _LOGGER.info(f"Key generated and stored at {domain_key_file}")
+            if self.data.key is None:
+                legacy_key_path = self.hass.config.path(DOMAIN_KEY_FILE)
 
-        self._crypto = None
+                if path.exists(legacy_key_path):
+                    with open(legacy_key_path, "rb") as file:
+                        self.data.key = file.read().decode("utf-8")
 
-        with open(domain_key_file, "rb") as file:
-            self._crypto = Fernet(file.read())
+                    remove(legacy_key_path)
+                else:
+                    self.data.key = Fernet.generate_key().decode("utf-8")
 
-    def encrypt(self, data: str):
-        encrypted = self._crypto.encrypt(data.encode()).decode()
+                await storage_manager.async_save_to_store(self.data)
+
+            self.crypto = Fernet(self.data.key.encode())
+
+    async def encrypt(self, data: str):
+        await self._load_key()
+
+        encrypted = self.crypto.encrypt(data.encode()).decode()
 
         return encrypted
 
-    def decrypt(self, data: str):
-        decrypted = data
+    async def decrypt(self, data: str):
+        await self._load_key()
 
-        if decrypted.endswith("="):
-            decrypted = self._crypto.decrypt(decrypted.encode()).decode()
-        else:
-            _LOGGER.warning(
-                "BlueIris Server password is not encrypted, please remove integration and reintegrate"
-            )
+        decrypted = self.crypto.decrypt(data.encode()).decode()
 
         return decrypted